+++ core/domain/interface/after-sales/refund.go
@@ -58,8 +58,8 @@ type (
 		// 编号
 		Id int32 `db:"id" pk:"yes" auto:"no"`
 		// 金额
-		Amount float32 `db:"amount"`
-		// 退款方式：1.退回余额  2: 原路退回
+		Amount int64 `db:"amount"`
+		// 退款方式：1.退回余额 2: 原路退回
 		//RefundType int `db:"refund_type"`
 		// 是否为全部退款
 		//AllRefund int `db:"all_refund"`
